+++ lua/github-theme/types/gt.lua
@@ -23,7 +23,7 @@ gt.HighlightStyle.Stanout = 'stanout'
 gt.HighlightStyle.NoCombine = 'nocombine'
 gt.HighlightStyle.Strikethrough = 'strikethrough'
 
----@alias gt.TerminalConf "'conf'" | "'yml'" | "'itermcolors'" | "'colorscheme'" | "'tmux'" | "'xresources'" | "windows_terminal"
+---@alias gt.TerminalConf "'conf'" | "'yml'" | "'itermcolors'" | "'colorscheme'" | "'tmux'" | "'xresources'" | "windows_terminal" | "foot"
 gt.TerminalConf = {}
 gt.TerminalConf.kitty = 'conf'
 gt.TerminalConf.alacritty = 'yml'
@@ -32,6 +32,7 @@ gt.TerminalConf.konsole = 'colorscheme'
 gt.TerminalConf.tmux = 'conf'
 gt.TerminalConf.xresources = 'Xresources'
 gt.TerminalConf.windows_terminal = 'json'
+gt.TerminalConf.foot = ''
 
 ---@class gt.Highlight
 ---Foreground Color